Move DbWhitelister into wagtail.admin.rich_text.converters.editor_html
