+++ wagtail/admin/rich_text/__init__.py
@@ -7,7 +7,8 @@ from django.utils.module_loading import import_string
 
 from wagtail.utils.widgets import WidgetWithScript
 from wagtail.admin.edit_handlers import RichTextFieldPanel
-from wagtail.core.rich_text import DbWhitelister, expand_db_html, features
+from wagtail.admin.rich_text.converters.editor_html import DbWhitelister
+from wagtail.core.rich_text import expand_db_html, features
 
 
 class HalloPlugin:
