+++ wagtail/core/tests/test_rich_text.py
@@ -3,8 +3,7 @@ from django.test import TestCase
 from mock import patch
 
 from wagtail.core.models import Page
-from wagtail.core.rich_text import (
-    DbWhitelister, RichText, expand_db_html)
+from wagtail.core.rich_text import RichText, expand_db_html
 from wagtail.core.rich_text.feature_registry import FeatureRegistry
 from wagtail.core.rich_text.pages import PageLinkHandler
 from wagtail.core.rich_text.rewriters import extract_attrs
@@ -41,43 +40,6 @@ class TestPageLinkHandler(TestCase):
     def test_expand_db_attributes_not_for_editor(self):
         result = PageLinkHandler.expand_db_attributes({'id': 1})
         self.assertEqual(result, '<a href="None">')
-
-
-class TestDbWhiteLister(TestCase):
-    def setUp(self):
-        self.whitelister = DbWhitelister()
-
-    def test_clean_tag_node_div(self):
-        soup = BeautifulSoup('<div>foo</div>', 'html5lib')
-        tag = soup.div
-        self.assertEqual(tag.name, 'div')
-        self.whitelister.clean_tag_node(soup, tag)
-        self.assertEqual(tag.name, 'p')
-
-    def test_clean_tag_node_with_data_embedtype(self):
-        soup = BeautifulSoup(
-            '<p><a data-embedtype="image" data-id=1 data-format="left" data-alt="bar" irrelevant="baz">foo</a></p>',
-            'html5lib'
-        )
-        tag = soup.p
-        self.whitelister.clean_tag_node(soup, tag)
-        self.assertEqual(str(tag),
-                         '<p><embed alt="bar" embedtype="image" format="left" id="1"/></p>')
-
-    def test_clean_tag_node_with_data_linktype(self):
-        soup = BeautifulSoup(
-            '<a data-linktype="document" data-id="1" irrelevant="baz">foo</a>',
-            'html5lib'
-        )
-        tag = soup.a
-        self.whitelister.clean_tag_node(soup, tag)
-        self.assertEqual(str(tag), '<a id="1" linktype="document">foo</a>')
-
-    def test_clean_tag_node(self):
-        soup = BeautifulSoup('<a irrelevant="baz">foo</a>', 'html5lib')
-        tag = soup.a
-        self.whitelister.clean_tag_node(soup, tag)
-        self.assertEqual(str(tag), '<a>foo</a>')
 
 
 class TestExtractAttrs(TestCase):
